requirements.txt
# Core RAG dependencies
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.0
chromadb>=0.4.0
numpy>=1.24.0
tiktoken>=0.5.0

# Document processing
pypdf>=3.17.0
python-docx>=1.0.0
markdown>=3.5.0

# CLI
click>=8.1.0
rich>=13.0.0

# Optional: For API support
fastapi>=0.104.0
uvicorn>=0.24.0
